requirements.txt
# Backend
flask>=3.0.0,<4.0.0

# Google Cloud - Versions compatibles
google-cloud-firestore>=2.16.0,<3.0.0
google-cloud-dialogflow>=2.26.0,<3.0.0

# Google AI
google-generativeai>=0.3.2,<1.0.0

# Google ADK (Agentic Development Kit)
# Note: ADK utilise google-generativeai en interne
google-genai>=0.1.0

# Utilitaires
python-dotenv>=1.0.0
pydantic>=2.0.0
gunicorn>=21.2.0


flask_cors>=4.0.0
requests>=2.31.0

# Pour language_v1
google-cloud-language>=2.10.0

Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7